requirements.txt
aniso8601==7.0.0
asgiref==3.5.0
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==3.2.12
django-debug-toolbar==3.2.4
django-filter==21.1
django-graphiql-debug-toolbar==0.2.0
django-graphql-jwt==0.3.4
django-heroku==0.3.1
flake8==4.0.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
iniconfig==1.1.1
mccabe==0.6.1
mypy==0.942
mypy-extensions==0.4.3
packaging==21.3
pluggy==1.0.0
promise==2.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.1.1
pytest-django==4.5.2
python-decouple==3.6
pytz==2022.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.1
whitenoise==6.0.0
